Automated default environment for python 2 & 3.
Implement all the glue necessary to provide the
automated creation of python virtual environments.
The intention is both to support a unit-level test
suite and also to ease project local setup burden.

These will include required python deps arranged
to allow seamlessly switching between versions.
This is done to aid both in development but also
support a unit-level test suite.

Add a Makefile with test targets wired for test
discovery and have these depend on targets which
do environment creation and installation of libs.
For Python 3 use the builtin venv + pip while for
python 2 we leverage docker and package the deps
inside a container in which tests then execute.
Things will only be rebuilt if changes are made
to requirements.txt.

Note that the selection of python 2 is done simply
by specifying a variable when invoking make which
switches the PYTHON_BIN being executed. A trivial
"test" is included to allow exercising this stuff.

Examples:

Python 3:
$ make test

Python 2:
$ make PY=2 test