Arrange for mig test cases to throw if any leaked resources are detected.
Integrate resource warnings with the FakeLogger - this involves registering
the fake logger as the stream for captured warnings to be delivered to. Add
some matching logic for "ResourceWarning: unclosed file ..." exception
messages, pull out the line number and output file, and record this stuff.
When the tests exits the teardown method checks whether any such records
exists and if so throws loudly including these details in a well formatted
message that should allow tracking down the cause.

The overarching goal here is so to ensure that any code being executed
under test can leave an open file and get away with it.