Arrange cross-python functioning generateconfs and cover its operation.
Consume the ensure_native_string() compat function to make the install
package and thus generateconfs functional on Python 3. This ends up
effectively being a backport of changes already done in experimental
but switches it to use the tested string compatibility function.

Include the most minimal set of tweaks to mig.shared.install to allow
some arguments to be supplied/overidden such that the output files can
be generated in a specified directory. Lean on keyword_auto as a hint
for cases where values (such as salt and other random numbers) are
generated which allows tests to pass in known values. Add a fixture
for expected output content and assert that generated files match this.

While here fix a small bug due to the use of os.path.exists that should
have been os.path.lexists due to dealing with a symlink - the check for
existince of the confs symlink would not behave correctly if it did not
point to a valid config folder. Fix this and cover it with a test thus
ensuring that the config generation recovers from a bad symlink.