Address filehandle leaking across file writing methods in fileio.
The existing code places a try/catch around both the write() and close()
rather than closing in a finally - the effect of that being that if the
write raised an exception close would not be called. This was noticed when
running the code under test produced ResourceWarning messages (something
that is now a hard error in the test support code now it matches these).

Fix this by jumping straight to a standard context manager syntax. Given
the issue existed for both chunk and whole file writing, and having noted
that those functions were nearly identical, refactor whole file writes in
terms of chunk writes thus it inherits the fixes. Cover all of this.