Bringup of the configuration object cross-version and coarse coverage.
Arrange environment specific test configurations such that codepaths
that depend upon it (i.e call the get_configuration_object() function)
can be used under test. Do this by providing an explicit config making
facility as part of the enviromnent helpers which operates based on an
environment name e.g. "local" and "test".

For cross-version support the test configuration must be adapted for
differences in paths (specifically Python 2 where execution occurs
within a container and thus container-relative paths are necessary).
Support a --python2 command line argument in makeconfig which causes
the paths adjusted as necessary.

Add a wrapper to utilise the pre-existing testcore unit tests as part
of the automatde test suite. This is done in order to exercise some
codepaths that depend on a valid configuration object. Note though
that this approach can likely be used as an examplar of the wrapping
necessary to expose existing tests.