Integrate autopep8 into the Makefile and set some baseline default options.
We are currently in a situation of wanting to make sure that out codebase
is autopep8 clean but there is no "standardised" way of ensuing this. Take
the opportunity to codify this in a new `make fmt` target - arrange the
autopep8 dependency as additional local dependency and invoke it via the
envhelp infrastructure.

As part of this commit we configure autopep8 with an option to ignore E402
which disables the re-arrangement of imports - this is relevant to us as
we carefully arrange import order and the current workaround (wrapping an
extra try-except around the imports with a comment that says its to trick
autopep8) is pretty messy in terms of the source code.