Ensure viable PYTHONPATH when executing environmentally configured python.
Doing so means that we support the ability to execute arbitrary scripts
within the tree directly via our wrappers - which one needs to do anyway
to have a full assortment of the requisite dependencies - wbile avoiding
the need to alter sys.path in non-test files.

While here drop an __init__.py file in the tests folder thus allowing
tests to be accessible as modules in the tests.<filename> and thus
allows easily specifying specific tests to be run with unitttest.