Create structured constants for validation dicts used for openid auth.
The validation of openid auth parameters is achieved by the use of a
generic validation routine which is passed differing dictionaries based
on which open id auth mode is in use.

Previously these dictionaries were buried cases within what amounted to
a switch over the auth mode. In addition we note the auth mode is an
enumerated set of values and that codepaths take action based on which
of the set is in use.

Make the auth mode an enumerated constant. Use the existing starting
values of the auth constants as the enum values which are themselves
defines as string subclasses such that they can be used interchangably
in cases where the originals were used to support gradual replacement.

Add the concept of named constants whose intent is to allow definition
of an an immutable key-value structure whose keys are a know set of
values. Use this to define and expose the default autocreate openid
validation dicts against their enum constant.