Split up skip_log into individual skip_X_log for the three common mig, auth and gdp logs. We used to only rely on skip_log for skipping mig.log init in daemons already doing their own logging to individual DAEMON.log files. It turns out it breaks auth.log if we interpret it as a general do not init logs at all as is the case since merging PR77.
On a related note we probably want to look into the case of skipping gdp log
init, which used to hinge on the same skip_log variable and thus would not be
done in the daemon init code, and with the split was toggled to False by
default. AFAICT gdp.log was in fact still handled from daemons despite the
skip. Likely because it was explicitly handled in griddaemons auth or session
code.

NOTE: this commit breaks the test_mig_unittest_testcore.MigUnittestTestcore
test because it with the new default then expects auth.log to be used even in
the test. That is work for a follow-up commit, if possible before merging.