Minimal changes to enforce safeinput assertions & repair email validation.
The wiring of the pre-existing safeinput tests into the automated suite
was executing them but did not raise assertion errors upon test strings
not being handled correctly thus did not enforce correctness.

Break the test values into pass validation/fail value lists. Loop these
lists and use a similar approach to testcore wherein a non-zero exit is
triggered for if a test string passes is allowed/rejected when it should
not be and wire these exits to trip AssertionError when under test.

In doing so it was discovered that email validation was not behaving the
same across versions and this turns out to be a combination of dependency
versioning problems in combination with overly broad assertion handling
in the suite masking issues. Address this here.