Introduce structured arg parsing and reimpelent generateconfs atop it.
Prior work made the defaults used for command line arguments in the
generateconfs utility accessible as a structure. Make use of this to
jettison our custom code and instead use all the user-facing niceties
of the argparse library.

This commit introduces a argument parsing wrapper for local use named
minimist in a nod to the node library of the same name from which we
largely borrow the externally facing API. It was remarkably suitable
for our purposes and a close match to our existing code which operates
in terms of arrays of argument names which must be treated a certain
way i.e. string, boolean, etc.