Apply Python 3 changes to grid_openid allowing test instantiation.
Make use of a need to cover the basic operation of this code for some
minimal changes that allow the server it implements to be trivial to
instantiate directly as an object. Use this ability to do so under test
and add the necessary wiring to issue a single top-level request from
the server and validate it responds with no internal issues logged.

Note that a top-level request was not previously succeedding due to
certain values being errantly dereferenced (though the code was clearly
written to have been safe under these circumstances). Fix this here.

---

align requirements with extant reality

ensure server is closed between tests and validate with a status 404 case

pull the test synchronisation up a level thus out of the server logic

ensure the server is fully managed inside test case arranged thread

improve naming

make the class configurable an extend documentation

inline escape_hmtl

fixup

fixup

fixup

fixup

remove unnecessary version swizzle

make use of the same Py3 compatible openid library as rocky 9

fixup