Follow-up to review from albu-diku. Thanks. Cleaned up unit tests with native asserts changed to MiGTestCase ones, MRSL strings moved to top-level constants and file writing switched to modern  `with open() as fd` pattern for much shorter tests.
Added a TODO in settings about showing the check_types function some 'love'. It
currently has a quite misleading name as it does not just check types but in
fact fills the given dictionary in-line. Either the name should reflect that or
it should be split into separate functions in the places where it's used.