Backport NativeStringIO wrapper helper from experimental to shared.compat and start using it in shared.ssh where we really need a platform-specific StringIO to write generated ssh keys to. The existing solution does NOT work on py3 and is a missing piece at least in the cloud integration and probably also in the jupyter integration there.
Migrate the old deprecated base64.decodestring in shared.ssh to use the modern
b64decode, which happens to also just work without string tweaks in this case.

Implemented unit tests for the paramiko key handling parts of shared.ssh
functions. The remaining ssh functions require an actual ssh service to test
against so they are left out of the unit tests for now.