PR104 basically exposed permanent_freeze for the cli, but handling the value as a boolean. The actual production use relies on the space-separated list of strings for selecting individual Archive flavors to write protect like:   --permanent_freeze="freeze backup"
Adjust the type of the arg from bool to string and simplify the handling as the
value is just passed as-is to the generated MiGserver.conf and effectively
parsed in shared.configuration on use.

Mangle the calling API a bit to enable the caller to easily work with the parsed
result in addition to the options dict.