Provide infrastructure for assertion of related values as a single case.
In out desired use of the test suite is a requirement to allow writing
a test that is the same assertion agianst a series of related values.
As an example, imagine an operation on chunks and wishing to assert
multiple sizes.

This is obviously something to be extremely careful in promoting given
it is in a fair amount of conflict with the notion of separate test
cases, which experience says more "generally" leads to better outcomes.
However, used judiciously this can be useful, and this is far better
supported by a mechanism designed to support it: namely, that _all_
failures are reported up front rather than the first failing assertion
cancelling the test block and the ending up in a fix-and-rerun cycle.

Add a facility and associated assertion to the test support library
that allows executing an operation and its assertion against a series
of values, buffers any exceptions that occurs and enforce that either
a no exception or all exception outcome is asserted by end of test.

Include a mechanism to the support library so arbitrary objects can do
checks and use this in multi-value assertions to ensure a result is
validated by its containing case.