Carve out the server testing facility as a separate support lib file.
This provides a support library facility for running http servers
under test in a separate thread - something that is necessary because
we need tests to be able to issue requests against them.

Note for the historical record: there was an attempt to make the
executon wrapped inherit from Thread but experience showed that it
was just too difficult to get right when also intending to keep a
reference to an event that must be consulted from the parent - opt
insteda to create an objectthat quacks much like a Thread but creates
one internally at a time of its pleasing.