Make createuser work for new installations and add basic test.
Allow user creation to provision the user_db_home directory as well as
the file itself which is the situation after config generation is run
to completion for a new installation.

Cover the very basic operation of createuser ensuring that a user db
and lock file are correctly create upon a call to create a test user.