Implement a test covering operation of the cat functionality file.
Make use of the seam that exists by virtue of the functionality files
return output objects which are interpreted by a wrapper to unit test
the behaviour of cat.

Specifically, we can arrange for a suitable environment plus arguments
and then check the right output_objects are created. Allow doing this
by exposing a variant of the cat main method which returns them rather
than invoking the outer wrapper. Name this _main and call it from the
original main.

Note that this change is in effect a blueprint that can be used when
adding coverage for other functionaity files.