Extend MigTestCase with such that it can provide a configuration object.
A great deal of code that we wish to exercise under test requires a
configuration object. The wiring to generate test configuration values
is already in tree, but actually consuming it in tests was done in an
ad-hoc fashion. Meanwhile the separate notion of a FakeConfiguration
was added primarily to allow exercising code that needs a specific
subset of vars set with specific values.

This patch raises this use-case to the level of the common support
library. Use the way a logger was exposed as a blueprint and add a
configuration property to MigTestCase - this is a computed property
that lazily creates a (by default) FakeConfiguration object.

This object persists for the duration of the test. Provide a means for
easily overriding this to make it a fully populated Configuration
instance using the testconfig - this mechanism is _declarative_ such
that the details of configuration loading are transparent to callers.
Document this use via an extra test case in the support library tests.