Adjust the recent wwwserve_max_bytes addition in cat.py to show a user friendly error in the web file manager (Files) when the limit is exceeded.
Reworked the backend and output handling a bit to deliver the size error as an
explicit HTTP "413 Payload Too Large" status code in line with
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status

Added a new downloadWrapper helper in file manager to handle binary data
downloads and react to errors by displaying them in the usual popup with an
extra inline hint in the status bar.

Serving binary files over AJAX is a bit cumbersome and apparently makes it
rather difficult to extract and show the plain text error message in case of
this exceeded max bytes case so a generic message is used for now.