Introduce argparse based argument parsing with getopt compatibility.
Add a shared module intended to handler arguments parsing. Expose
methods that wrap argparse accepting list of named arguments to treat
as strings, booleans and integers respectively. This is purposefully
designed as drop-in for the style used in the generateconfs cli.

Also implement a getopt replacement. This parses getopt strings in
order to build up known arguments then calls the previously defined
shared module - in this way the new module is also a drop-in for any
callsite using getopt (e.g. createuser).

Support both the naming of arguments and help text. Additionally
provide a function able to break up pre-existing usage strings used
with relatively few changes can be re-used for new-style arguments.