Implement declarative argument definitions and logic to bundle values.
Expand the arguments module with the notions of a defined grouping of arguments
(ArgumentBundleDefinition) and bundles of particular arguments (ArgumentBundle).

Use this as the mechanism by which payloads are checked for validity. As part
of declaring a bundle definition the expected positional arguments are declared
so implement basic length checks that catch missing positional arguments which
were required. Each argument itself is also tested aganst an optional validity
function which can be specified at the point of definition.