Initial version of standalone account user page with site registered account details, OIDC token info if available and basic account management: reset password, renew account and delete account.
Delete account is only a skeleton based on pwreset so far and just sends out an
email to registered address with a similarly authenticated link to use for the
confirmation of the complete account removal with all data. The remaining parts
and operator actions to match such requests are still not implemented and the
Delete button is therefore disabled with a note about contacting support.