Add support for expansion of external sources for the sftp , davs and ftps key/cert fingerprints in MiGserver.conf in order to get rid of tedious recurring manual fingerprint updates when relying on LetsEncrypt certificates, which come with frequent renewal.
Allows the user_PROTO_key_sha256 configuration variables to e.g. be set to
FILE::/path/to/fingerprint
in order to always read the current fingerprint from that file.
It also supports the usual
FILE::/path/to/fingerprint$$/path/to/fast/cache/fingerprint
for automatic caching in a fast memory-backed or similar cache location.

Adjust the migcheckssl cron job helper to inform about it and automatically
write the combined pem and pub sha256 fingerprints in the corresponding
'combined.pem.sha256' and 'combined.pub.sha256' files for complete automation
of the flow in that case.

The sftp key remains unchanged in that setup so one can keep that fingerprint
static in MiGserver.conf, or use that combined.pub.sha256 if changes are ever
expected. Such changes are typically rather inconvenient for clients as they
have to find and update their local known hosts file unless also delivered e.g.
on DNSSEC.

Updated unit test templates to fit migcheckssl changes.