Repair the display of error messages in the 'unhappy' output path.
Evidence suggests that at some point the output rendering path became
sensitive to the presence and ordering of objects - the result being that
the objects that are assigned in the case of displaying an error were not
sufficient to trigger the generation of a valid HTML page.

There is likely a deeper issue here, but opt to fix the immediate problem
by assigning an appropriate list of output objects that result in a valid
HTML pages being generated with the error object rendered within the body.