Introduce support for jinja2-based templates.
This adds the necessary logic both to maintain a store of templates
and making them available for rendering within any page via a new
template output object type.

The template type is wired into the output path. The existing path for
object validation check the arguments passed to the template only, so
so a specific addition is made to the validation code to ensure validity
of the template object itself. In the case of a requested template being
missing a specific exception is thrown. Logic is also added to identify
the specific situation of a missing cache directory and report that as
its own error to aid operators in tracking down any issues.

Make sure there is a separation of the test templates from the mainline
stuff by wiring a TEMPLATES section into Configuration and reading the
necessary paths via the loaded configuration. Doing so allows the cache
directory as well as top-level or "base" packages from which to read
templates to be specified.

This configuration facility is used within the tests to adjust the
paths and the section is exposed to the generateconfs command line
similarly to other values. The default location for the template cache
is a directory within the state folder which was chosen so it can be
mounted over by docker-migrid as is done for other temporary paths.

In the process of doing this work a test was added to checkconfs due to
some changes being nearby and in doing so discovering that it would not
successfully complete against an empty file. So checkconfs is fixed here
also to ensure the coverage of all code affecting the use of templates.