Make create_user work for new installations and add basic test.
Allow user creation to provision the directries necessary for user
creation to succeed in addition to the database file itself given the
state that exists after config generation is run to completion for a
new installation.

Cover the very basic operation of creat_user ensuring that a user db
is created (along with its lock file) and that a user entry is added
to the database.