Make create_user work for new installations and add basic test.
Allow user creation to provision the directories necessary for user
creation to succeed in addition to the database file itself against
the state as it is after config generation is run to completion for a
new installation.

Cover the very basic operation of create_user ensuring that a user db
(along with its lock file) will be created and a user entry added to it.