Remove install createuser function and the dev acct script dependent on it.
There has ben a better alternative to this in the form of the create_user
function from useradm and the mig/server/createuser script which is usable
from the command line. In the case of the latter, the addition of the
testconfigs and supporting infrastructure mean that all pre-requisites
which allow use of the newer replacement are provided.

Remove this and in so doing cut a dependency on the removed crypt module.