Add test support code that allows capturing output snapshots & an example.
A large surface area of code exists that generates e.g. HTML output. Much
of this is not built upon names types and there is often no easily usable
seam that exists for capturing this structure and/or supplying a fixture.

This change lays some groundwork that is intended to allow such code to
be refactored going forward without the fear of introducing regressions.
Roughly, the idea is to first write a test which captures the output as-is
given suitable inputs. The snapshotting code will do this automatically on
first invocation. These can then be checked in as test-only changes. Once
such a test is in-place future execution of the test will assert that the
output generated matches the previously saved snapshot.

Snapshots can be refreshed either by deleting their corresponding on-disk
file or specifying true, yes or 1 for the REFRESH_SNAPHOTS env variable.