Implement declarative payloads.
Add the logic to automatically package values as a bundle that is formally
named. Such a bundle remembers its type, can be validated and provides the
facilities allowing its serialisation and deserialisation.

In terms of utility, such a construct allows us to capture a particular
grouping of values that belong together and is intended as infrastructure
for cases where indirect calling is necessary e.g. programmatic accesss to
a JSON API where a client and a server API must exchange strucrured data.

The conditions needed to declare a payload "valid" live with its definition
thus any set of values can be immediately tested for validity. The code
used to perform the checks is is the existing validation functions, thus
there is an upwards path for adoption in existing code.

Use this as the mechanism by which payloads are checked for validity. As part
of declaring a bundle definition the expected positional arguments are declared
so implement basic length checks that catch missing positional arguments which
were required. Each argument itself is also tested aganst an optional validity
function which can be specified at the point of definition.

Raise uniform payload exceptions when validation fails. Further, upon
failure the exception produced can be serialised directly to JSON with
useful messages about the failures that occurred, making it a useful both
for the case of generating detailed user-facing reports on the command
line as well as potentially reporting errors against fields with UIs.