Add support for adding request contextual data to configuration objects.
This change adds a small namespace to configuraton objects that behaves
like a dictonary allowing objects to made available to e.g. request
handlers.

Opt to do this by specialising Configuration rather than changing it
directly. The rationale is to preserve a distinction between static
parts of the configuration of a dpeloyment, which do no change at
runtime, versus e.g. a users locale which would be different for
each request though consistent for its duration.

Such functionality it needed to support templates without restorting
to globals within the current system architecture; only congifuration
objects are already threaded through most of the places this would
need to be available with the correct semantics of an instance being
created wherever it is needed.