Lay groundwork for splitting runtime and static configuration data.
This change paves the way adding further runtime specific data to
configuration objects. An instance of a configuration already contains
such data, such as the mig_server_id property when in use by daemons
and a reference to the logger. With needs arising for further runtime
data, make a clear separation between runtime attributes and the
static configuration data clear.

Opt to do this by making an object which internally keeps a reference
to an underlying configuration object, proxies attribute access to it
and can then be used as a drop-in substitute in ay call stack of logic.
Name this RuntimeConfiguration.

Stop short of moving existing runtime properties as of this commit.