Arrange the requisite explicit environ to the tested functionality files.
At some point the couple of tests we have for functionality files started
to fail. The way these tests were achieved was a minimal change was made
to the mainline code itself to allow passing in a controlled series of
arguments, which the test file then uses to exercise the codepaths.

Make sure this works fully across cat and datatransfer, with the following
being required:
* in datatransfer, the populated environment was not being passed into the
  certificate validate method - do so thereby aligning it with cat
* the tests for both cat and datatransfer were not passing the test
  populated environment in all cases - unclear exactly how that worked
  previously (sense if that we added or tightened a check somewhere) but
  get this value passed down