Double down on human readable fixtures and support via extended hints.
One of the key aspects of using fixtures is maintaining the ability to
know their contents with ease. This was slightly undermined by allowing
pickled fixture data.

Use the opportunity presented by having to handle the inclusion of
timestamps at now points in the future, something that is something
of a necessity for use of fixtures going forward, both to add support
for doing so but push it further so such values are specified within
a file readable format.

There were pre-existing provisions for loading JSON as well as being
able to specify additional transformations of properties to perform
as a series of hints. Extend and make use of this mechanism for a new
fixture output path - this allows fixture data to be written out in a
forat different than it is read in (which can include transformation).

Make use of this new support to rework the test user provisioning. The
use is now defined in JSON and, rather than copy a pre-existing file
representing the requisite user database, write this database based
on the JSON data. Do the work necessary to make sure the data in this
file is output with strings transformed into bytes as is the format
of the user database.