Adjust `pwcrypto` slightly to support external use of the existing self-tests, like it's the case e.g. for `userio`, and integrate running it in unit testing.
Added a lot of unit tests to cover almost all functions in `pwcrypto` and to
include more tests of fundamental functionality against fixed known values as
suggested in PR343 discussion.
A few tests will require API changes to pass a random seed to allow that and
have been left as disabled for now with TODO comments added.