Expand hints with better pickle bits and use it to cover create_user.
Provide a means to load pickled data from disk to which the hints
mechanism as introduced for fixtures can be applied. Extensively
document the rationale for hinting in light of expansion of this
facility and its re-use.

Use all the facilities for a test covering useradm create_user function.
Note, create_user is mostly about writing rather than the population
of the user dictionary with a majority of the important prep work with
regard to what will be written being done by the calling script. This
means that the we are only comparing written data as of now, meaning it
is something of a loose test, but this towards additional rigour.

Note that we used the opportunity to make sure that calling create_user
writes exactly what was declared within our on-disk user fixture. This
begins to verify the fixture itself and strenghtens the assurances of
any test consuming that fixture (to the extent allowed given the
responsibilities of create_user noted above - a start).